LayoutTests/webaudio/OfflineAudioContext/offlineaudiocontext-can-only-render-once-expected.txt

OfflineAudioContext can only be started once.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS OfflineAudioContext threw exception when trying to start it again: InvalidStateError: Rendering was already started
PASS successfullyParsed is true

TEST COMPLETE
